Restructured the whole package
rename package to SIPH
added submodule models to include functions and submodules for various
projects
updated scripts and submodules for new structure
