+++ README.md
@@ -1,2 +1,2 @@
-# VNRT
-Near real-time monitoring of forest disturbance with VIIRS data
+# SIPH
+Satellite Image Processing Hub
