--- hls/__init__.py
@@ -1,14 +1,0 @@
-""" Module for functions for HLS
-"""
-from .hls import hls2stack, hlsQA, hn2ln
-from .mask import mn2ln, bit2mask, mask2array
-
-
-__all__ = [
-    'hn2ln',
-    'hls2stack',
-    'hlsQA',
-    'mn2ln',
-    'bit2mask',
-    'mask2array'
-]
