+++ models/__init__.py
@@ -1,0 +1,2 @@
+""" Modules for different models
+"""
