+++ models/hls/__init__.py
@@ -1,0 +1,2 @@
+""" Module for functions for HLS
+"""
