+++ scripts/download.sh
@@ -88,9 +88,9 @@ done
 # run python script
 cd /projectnb/landsat/users/xjtang/documents/
 if [ $method = "ftp" ]; then
-	python -m VNRT.tools.ftp_download -s $sensor -c $collection -p $product -t $h $v -y $year -d $dstart $dend $des
+	python -m SIPH.tools.ftp_download -s $sensor -c $collection -p $product -t $h $v -y $year -d $dstart $dend $des
 elif [ $method = "http" ]; then
-	python -m VNRT.tools.http_download -u $username -w $password -s $sensor -c $collection -p $product -t $h $v -y $year -d $dstart $dend $des
+	python -m SIPH.tools.http_download -u $username -w $password -s $sensor -c $collection -p $product -t $h $v -y $year -d $dstart $dend $des
 else
 	echo 'unknown method'
 fi
