+++ scripts/mask2stack.sh
@@ -52,5 +52,5 @@ done
 echo 'Total jobs to submit is' $njob
 for i in $(seq 1 $njob); do
     echo 'Submitting job no.' $i 'out of' $njob
-    qsub -j y -N HLS_$i -V -b y cd /projectnb/landsat/users/xjtang/documents/';' python -m VNRT.hls.mask_to_stack ${overwrite}${recursive}-p $pattern -b $i $njob -m $mask $ori $des
+    qsub -j y -N HLS_$i -V -b y cd /projectnb/landsat/users/xjtang/documents/';' python -m SIPH.models.hls.mask_to_stack ${overwrite}${recursive}-p $pattern -b $i $njob -m $mask $ori $des
 done
